agent-cockpit/requirements.txt
# InsightSage Agent - Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pyodbc==5.2.0
pyyaml==6.0.2
pydantic==2.12.5
pydantic-settings==2.7.1
httpx==0.28.1
apscheduler==3.11.0
cryptography==44.0.0
python-dotenv==1.0.1
loguru==0.7.3
slowapi==0.1.9
python-socketio[asyncio_client]==5.11.0

# Pour générer l'exécutable Windows
pyinstaller==6.19.0
